fix: bug item that turns into statues

+++ schema.sql
@@ -603,7 +603,7 @@ CREATE TABLE `player_depotitems` (
   `sid` int(11) NOT NULL COMMENT 'any given range eg 0-100 will be reserved for depot lockers and all > 100 will be then normal items inside depots',
   `pid` int(11) NOT NULL DEFAULT '0',
   `itemtype` int(10) NOT NULL,
-  `count` smallint(5) NOT NULL DEFAULT '0',
+  `count` int(11) NOT NULL DEFAULT '0',
   `attributes` blob NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
@@ -631,7 +631,7 @@ CREATE TABLE `player_inboxitems` (
   `sid` int(11) NOT NULL,
   `pid` int(11) NOT NULL DEFAULT '0',
   `itemtype` int(10) NOT NULL,
-  `count` smallint(5) NOT NULL DEFAULT '0',
+  `count` int(11) NOT NULL DEFAULT '0',
   `attributes` blob NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
@@ -646,7 +646,7 @@ CREATE TABLE `player_items` (
   `pid` int(11) NOT NULL DEFAULT '0',
   `sid` int(11) NOT NULL DEFAULT '0',
   `itemtype` int(10) NOT NULL DEFAULT '0',
-  `count` smallint(5) NOT NULL DEFAULT '0',
+  `count` int(11) NOT NULL DEFAULT '0',
   `attributes` blob NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
@@ -730,7 +730,7 @@ CREATE TABLE `player_rewards` (
   `sid` int(11) NOT NULL,
   `pid` int(11) NOT NULL DEFAULT '0',
   `itemtype` int(10) NOT NULL,
-  `count` smallint(5) NOT NULL DEFAULT '0',
+  `count` int(11) NOT NULL DEFAULT '0',
   `attributes` blob NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
